+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Changelog
   in regions where they may be required, and are not generally recommended.
 * We now ship ``manylinux_2_24`` wheels, in addition to our ``manylinux2010``
   and ``manylinux2014`` wheels.
+* Added ``rfc4514_attribute_name`` attribute to
+  :attr:`x509.Name <cryptography.x509.Name.rfc4514_attribute_name>`,
 
 .. _v3-4-7: